{z3,z3c}: Fix filename of tfa98xx preset files
Fix ".reset" to ".preset".
During boot tfa9890_amp module loads the preset files for different usecases. With the typo in one of the filenames, function tfa9890_prepare_for_ioctl will exit with error. Subsequent files will then not be loaded, especially the preset file for "VoiceCallEarpiece", resulting in distorted sound and volume,

Tested with LineageOS16 build 2020-02-09 for Z3: This fixes the in-call audio issues.

Change-Id: I603d083fd3ce964bdf58544dcb68588306c74311